% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanFunctions.R
\name{RemoveNA}
\alias{RemoveNA}
\title{These are functions created to clean the data set
featured in this package. These functions can be used on any data set.
Remove NAs}
\usage{
RemoveNA(X, x)
}
\description{
These are functions created to clean the data set
featured in this package. These functions can be used on any data set.
Remove NAs
}
